+++ core/src/lib.rs
@@ -1,10 +1,10 @@
 pub mod momentum;
-pub mod trend;
+pub mod overlap;
 pub mod utils;
 pub mod volatility;
 
 // Re-export all the modules
 pub use momentum::*;
-pub use trend::*;
+pub use overlap::*;
 pub use utils::*;
 pub use volatility::*;
