+++ core/src/overlap/sma.rs
